OneLab/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
boto3==1.34.34
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
hubspot-api-client==8.2.0
groq==0.4.2
python-linkedin==4.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
beautifulsoup4==4.12.3
lxml==5.1.0
aiosmtplib==3.0.1
email-validator==2.1.0.post1
streamlit==1.32.0 